# Food-Delivery/api_gateway_service/docs/swagger.yaml
basePath: /
definitions:
...
  courier.CourierLocationResponse:
    properties:
      courier_id:
        type: string
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      updated_at:
        description: ISO 8601 timestamp format (e.g., "2023-12-31T12:34:56Z")
        type: string
    type: object
...
  courier.CreateCourierLocationRequest:
    properties:
      courier_id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
    type: object
...
  courier.DeleteCourierLocationResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  courier.GetAllCourierLocationsResponse:
    properties:
      locations:
        items:
          $ref: '#/definitions/courier.CourierLocationResponse'
        type: array
    type: object
...
  courier.UpdateCourierLocationRequest:
    properties:
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
    type: object
...
  notification.CourierNotification:
    properties:
      courier_id:
        description: Bildirishnomani oladigan kuryer ID'si
        type: string
      created_at:
        description: Bildirishnoma yaratilgan vaqt
        type: string
      is_read:
        description: Kuryer bildirishnomani o'qiganligini bildiradi
        type: boolean
      message:
        description: Bildirishnoma matni
        type: string
      notification_id:
        description: Bildirishnoma ID'si
        type: string
      order_id:
        description: Bildirishnomaga bog'langan buyurtma ID'si
        type: string
      type:
        allOf:
        - $ref: '#/definitions/notification.CourierNotificationType'
        description: Bildirishnoma turi
    type: object
  notification.CourierNotificationResponse:
    properties:
      message:
        description: Javob xabari
        type: string
      notification:
        allOf:
        - $ref: '#/definitions/notification.CourierNotification'
        description: Kuryer bildirishnomasi ma'lumotlari (Create operatsiyasi uchun)
      success:
        description: Amaliyot muvaffaqiyatli bo'lganligini bildiruvchi flag
        type: boolean
    type: object
  notification.CourierNotificationType:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-comments:
      CourierNotificationType_NEW_RECOMMENDATION: Yangi tavsiya bildirishnomasi
      CourierNotificationType_ORDER_ASSIGNMENT: Buyurtma topshiriq bildirishnomasi
      CourierNotificationType_STATUS_UPDATE: Holatni yangilash bildirishnomasi
    x-enum-varnames:
    - CourierNotificationType_NEW_RECOMMENDATION
    - CourierNotificationType_ORDER_ASSIGNMENT
    - CourierNotificationType_STATUS_UPDATE
...
  notification.CreateCourierNotificationRequest:
    properties:
      courier_id:
        description: Bildirishnomani oladigan kuryer ID'si
        type: string
      message:
        description: Bildirishnoma matni
        type: string
      order_id:
        description: Bildirishnomaga bog'langan buyurtma ID'si
        type: string
      type:
        allOf:
        - $ref: '#/definitions/notification.CourierNotificationType'
        description: Bildirishnoma turi
    type: object
...
info:
  contact: {}
  description: API for Food Delivery System resources
  title: Food Delivery System API
  version: "1.0"
paths:
...
  /api/courier-locations:
    get:
      description: Retrieve a list of all courier locations in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/courier.GetAllCourierLocationsResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all courier locations
      tags:
      - Courier
  /api/courier-locations/{locationId}:
    delete:
      description: Remove a courier location from the system by its ID
      parameters:
      - description: Courier Location ID
        in: path
        name: locationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/courier.DeleteCourierLocationResponse'
        "400":
          description: Invalid Location ID
          schema:
            type: string
        "404":
          description: Location Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete courier location by ID
      tags:
      - Courier
    get:
      description: Retrieve details of a courier location by its ID
      parameters:
      - description: Courier Location ID
        in: path
        name: locationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/courier.CourierLocationResponse'
        "400":
          description: Invalid Location ID
          schema:
            type: string
        "404":
          description: Location Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get courier location by ID
      tags:
      - Courier
    put:
      consumes:
      - application/json
      description: Update details of an existing courier location by its ID
      parameters:
      - description: Courier Location ID
        in: path
        name: locationId
        required: true
        type: string
      - description: Courier Location Data
        in: body
        name: location
        required: true
        schema:
          $ref: '#/definitions/courier.UpdateCourierLocationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/courier.CourierLocationResponse'
        "400":
          description: Invalid Location ID
          schema:
            type: string
        "404":
          description: Location Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update courier location by ID
      tags:
      - Courier
  /api/courier-locations/create:
    post:
      consumes:
      - application/json
      description: Add a new courier location to the system
      parameters:
      - description: Courier Location Data
        in: body
        name: location
        required: true
        schema:
          $ref: '#/definitions/courier.CreateCourierLocationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/courier.CourierLocationResponse'
        "400":
          description: Invalid Data
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new courier location
      tags:
      - Courier
...
  /api/notifications/courier:
    get:
      description: Retrieve a list of all courier notifications
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notification.CourierNotificationResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List all courier notifications
      tags:
      - Notifications
    post:
      consumes:
      - application/json
      description: Add a new courier notification to the system
      parameters:
      - description: Notification Data
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/notification.CreateCourierNotificationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/notification.CourierNotificationResponse'
        "400":
          description: Invalid Data
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new courier notification
      tags:
      - Notifications
  /api/notifications/courier/{notificationId}:
    delete:
      description: Remove a courier notification from the system by its ID
      parameters:
      - description: Notification ID
        in: path
        name: notificationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Invalid Notification ID
          schema:
            type: string
        "404":
          description: Notification Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a courier notification
      tags:
      - Notifications
    get:
      description: Retrieve details of a courier notification by its ID
      parameters:
      - description: Notification ID
        in: path
        name: notificationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notification.CourierNotificationResponse'
        "400":
          description: Invalid Notification ID
          schema:
            type: string
        "404":
          description: Notification Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get courier notification by ID
      tags:
      - Notifications
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"